Master File Upload
This is a mostly empty unity folder. Copy and paste the things that you
need to put in here in their respective folders. Please do NOT delete
things from the file unless you are for SURE that they are not needed.
Also, be SURE that what you are adding works, or else you might mess up
the file.